+++ build.gradle
@@ -13,7 +13,9 @@ repositories {
 dependencies {
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.5.0'
     testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.5.0'
+}
 
+dependencies {
     String javaFxVersion = '11'
 
     implementation group: 'org.openjfx', name: 'javafx-base', version: javaFxVersion, classifier: 'win'
@@ -45,11 +47,11 @@ test {
 }
 
 application {
-    mainClassName = "Sados"
+    mainClassName = "head.Launcher"
 }
 
 shadowJar {
-    archiveBaseName = "Sados"
+    archiveBaseName = "SaDOS"
     archiveClassifier = null
 }
