+++ src/main/java/head/Sados.java
@@ -6,7 +6,6 @@ import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 
-import addon.Popup;
 import addon.Storage;
 import addon.Task;
 import addon.Task.Deadline;
